+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## 2.40.0 (2026-01-26)
 
 - added: Added change-server subscription timeout fallback.
+- added: Remove empty transaction metadata files when loading transaction data.
+- changed: Use syncRepo from upgraded edge-sync-client for syncing repo algorithm.
 
 ## 2.39.0 (2026-01-16)